Create the same number of entries for stubs, stub_helpers and la_symbol_ptr
Previously, we optimize lazily-bound symbols so that if the same symbol is
in GOT, we don't create an entry for that symbol in stub_helpers and
la_symbol_ptr. Even though this works as intended in most cases, it broke
some tools such as Facebook's fishhook that assume that stubs, stub_helpers
and la_symbol_ptr are parallel array.

Fixes https://github.com/bluewhalesystems/sold/issues/14