Emit source filename to a debug symbol
We have observed that lldb is sluggish when stopping at a breakpoint
if the binary is created with sold.

The obvious difference in the debug info section compared to lld/ld64
is that binaries created by sold have dummy source filename debug symbols.
I'm not sure if it's the cause of the issue, but it's worth to try to
set the same value as lld/ld64 to see if it'd fix the problem.

https://github.com/bluewhalesystems/sold/issues/29